+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+- Adicionado a função convert base64 to image.
+## 1.0.2
+- Adicionado a função getCRC16.
+
 ## 1.0.1
 
 - Adicionado a função retiraAcentos.